# docs/api/api.yaml
openapi: 3.0.0
info:
  title: Wutsi Earning API
  description: |
    API for managing blogger earnings
  version: 1.0.0
servers:
  - url: 'https://wutsi-earning-test.herokuapp.com'
    description: Sandbox
  - url: 'https://wutsi-earning-prod.herokuapp.com'
    description: Production
paths:
  '/v1/earnings/users/{user-id}':
    get:
      operationId: userEarnings
      parameters:
        - in: path
          name: user-id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: year
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchEarningResponse'
      security:
        - api_key:
            - earning
      tags:
        - Earnings
  /v1/earnings/compute:
    get:
      operationId: compute
      parameters:
        - in: query
          name: year
          required: true
          schema:
            type: integer
            format: int32
        - in: query
          name: month
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
      security:
        - api_key:
            - earning
      tags:
        - Earnings
  /v1/earnings/replay:
    get:
      operationId: replay
      parameters:
        - in: query
          name: start-date
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: end-date
          schema:
            type: string
            format: date
            nullable: true
      responses:
        '200':
          description: Success
      security:
        - api_key:
            - earning
      tags:
        - Earnings
components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    SearchEarningResponse:
      type: object
      properties:
        earnings:
          type: array
          items:
            $ref: '#/components/schemas/Earning'
    Earning:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        contractId:
          type: integer
          format: int64
          nullable: true
        partnerId:
          type: integer
          format: int64
          nullable: true
        date:
          type: string
          format: date
        amount:
          type: number
          format: int64
        currency:
          type: string
        syncDateTime:
          type: string
          format: date-time